refactor: move `findTestFiles` to a common directory where it can be reused in multiple builders
